requirements.txt
apache-airflow[celery]==2.10.3 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.10.3/constraints-3.12.txt
apache-airflow-providers-amazon
apache-airflow-providers-postgres[amazon]
numpy
pandas
pyarrow